+++ atlite/pv/orientation.py
@@ -7,7 +7,7 @@
 import sys
 import numpy as np
 import xarray as xr
-
+from numpy import sin, cos, deg2rad
 
 def get_orientation(name, **params):
     '''
@@ -51,13 +51,13 @@ def make_latitude_optimal():
 
         slope = np.empty_like(lat.values)
 
-        below_25 = lat.values <= np.deg2rad(25)
-        below_50 = lat.values <= np.deg2rad(50)
+        below_25 = lat.values <= deg2rad(25)
+        below_50 = lat.values <= deg2rad(50)
 
         slope[below_25] = 0.87 * lat.values[below_25]
         slope[~below_25 & below_50] = 0.76 * \
-            lat.values[~below_25 & below_50] + np.deg2rad(0.31)
-        slope[~below_50] = np.deg2rad(40.)
+            lat.values[~below_25 & below_50] + deg2rad(0.31)
+        slope[~below_50] = deg2rad(40.)
 
         return dict(slope=xr.DataArray(slope, coords=lat.coords), azimuth=180.)
 
@@ -65,8 +65,8 @@ def make_latitude_optimal():
 
 
 def make_constant(slope, azimuth):
-    slope = np.deg2rad(slope)
-    azimuth = np.deg2rad(azimuth)
+    slope = deg2rad(slope)
+    azimuth = deg2rad(azimuth)
 
     def constant(lon, lat, solar_position):
         return dict(slope=slope, azimuth=azimuth)
@@ -89,8 +89,8 @@ def SurfaceOrientation(ds, solar_position, orientation):
     vector analysis, Renewable Energy, 32(7), 1187–1205 (2007).
     """
 
-    lon = np.deg2rad(ds['lon'])
-    lat = np.deg2rad(ds['lat'])
+    lon = deg2rad(ds['lon'])
+    lat = deg2rad(ds['lat'])
 
     orientation = orientation(lon, lat, solar_position)
     surface_slope = orientation['slope']
@@ -99,14 +99,14 @@ def SurfaceOrientation(ds, solar_position, orientation):
     sun_altitude = solar_position['altitude']
     sun_azimuth = solar_position['azimuth']
 
-    cosincidence = (np.sin(surface_slope) * np.cos(sun_altitude) *
-                    np.cos(surface_azimuth - sun_azimuth) +
-                    np.cos(surface_slope) * np.sin(sun_altitude))
+    cosincidence = (sin(surface_slope) * cos(sun_altitude) *
+                    cos(surface_azimuth - sun_azimuth) +
+                    cos(surface_slope) * sin(sun_altitude))
 
     # fixup incidence angle: if the panel is badly oriented and the sun shines
     # on the back of the panel (incidence angle > 90degree), the irradiation
     # would be negative instead of 0; this is prevented here.
-    cosincidence.values[cosincidence.values < 0.] = 0.
+    cosincidence = cosincidence.clip(min=0)
 
     return xr.Dataset({'cosincidence': cosincidence,
                        'slope': surface_slope,